fix: address identity tag review comments
Extract shared identity tag UI helpers, align identity SELECT list formatting, and clean up tagging documentation wording. Keep partial tag submission behavior unchanged so incomplete rows still surface backend validation instead of being silently dropped.
