+++ crates/admin-ui/web/src/routes/identity/users.tsx
@@ -46,6 +46,7 @@ import {
   resendIdentityUserPasswordInvite,
   updateIdentityUser,
 } from '@/server/admin-data.functions'
+import { EntityTagBadges, EntityTagsField, sanitizeEntityTags } from '@/routes/identity/-entity-tags'
 import type {
   CreateUserInput,
   CreateUserResult,
@@ -1113,82 +1114,6 @@ function sanitizeUpdateForm(
   }
 
   return update
-}
-
-function EntityTagsField({
-  label,
-  tags,
-  onChange,
-}: {
-  label: string
-  tags: CreateUserInput['tags']
-  onChange: (tags: CreateUserInput['tags']) => void
-}) {
-  function updateTag(index: number, field: 'key' | 'value', value: string) {
-    onChange(tags.map((tag, tagIndex) => (tagIndex === index ? { ...tag, [field]: value } : tag)))
-  }
-
-  return (
-    <Field>
-      <FieldLabel>{label}</FieldLabel>
-      <div className="flex flex-col gap-2">
-        {tags.map((tag, index) => (
-          <div key={index} className="grid gap-2 sm:grid-cols-[1fr_1fr_auto]">
-            <Input
-              value={tag.key}
-              onChange={(event) => updateTag(index, 'key', event.target.value)}
-              placeholder="cost-center"
-            />
-            <Input
-              value={tag.value}
-              onChange={(event) => updateTag(index, 'value', event.target.value)}
-              placeholder="platform"
-            />
-            <Button
-              type="button"
-              variant="ghost"
-              onClick={() => onChange(tags.filter((_, tagIndex) => tagIndex !== index))}
-            >
-              Remove
-            </Button>
-          </div>
-        ))}
-        <Button
-          type="button"
-          variant="secondary"
-          onClick={() => onChange([...tags, { key: '', value: '' }])}
-          disabled={tags.length >= 5}
-        >
-          Add tag
-        </Button>
-      </div>
-      <FieldDescription>
-        Up to five lowercase key/value tags. Keys may use letters, digits, dot, underscore, or dash.
-      </FieldDescription>
-    </Field>
-  )
-}
-
-function EntityTagBadges({ tags }: { tags: UserView['tags'] }) {
-  if (tags.length === 0) {
-    return <span className="text-xs text-[var(--color-text-soft)]">No tags</span>
-  }
-
-  return (
-    <div className="flex flex-wrap gap-2">
-      {tags.map((tag) => (
-        <Badge key={tag.key} variant="secondary">
-          {tag.key}:{tag.value}
-        </Badge>
-      ))}
-    </div>
-  )
-}
-
-function sanitizeEntityTags(tags: CreateUserInput['tags']) {
-  return tags
-    .map((tag) => ({ key: tag.key.trim(), value: tag.value.trim() }))
-    .filter((tag) => tag.key.length > 0 || tag.value.length > 0)
 }
 
 function isOidcDisabled(form: CreateUserInput, providers: IdentityUsersPayload['oidc_providers']) {
